requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
streamlit==1.39.0

pandas==2.2.3
numpy==2.1.2
requests==2.32.3
python-dotenv==1.0.1

yfinance==0.2.44
# marketstack-api optional, comment out if error
# marketstack-api==0.1.3

sentence-transformers==3.1.1
chromadb==0.5.11

torch==2.4.1
scikit-learn==1.5.2

langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.8

plotly==5.24.1
